+++ packages/dsw-document-worker/CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Support item title for all types of questions
+
 
 ## [4.19.1]