+++ packages/dsw-document-worker/CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+
+## [4.20.0]
+
 ### Added
 
 - Support item title for all types of questions
@@ -15,7 +18,6 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 - Fixed file replies linked from file questions
-
 
 ## [4.19.1]
 
@@ -449,3 +451,4 @@ Released for version consistency with other DSW tools.
 [4.18.3]: /../../tree/v4.18.3
 [4.19.0]: /../../tree/v4.19.0
 [4.19.1]: /../../tree/v4.19.1
+[4.20.0]: /../../tree/v4.20.0